requirements.txt
Pillow==10.4.0
requests==2.32.3
numpy==1.24.4
transformers==4.46.3
scikit-learn==1.3.2
datasets==3.1.0
streamlit==1.40.1
matplotlib==3.7.5
scikit-image==0.21.0
torch==2.4.1
torchvision==0.19.1
altair==5.4.1
evaluate==0.4.3
accelerate==1.0.1
